% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pleiotropy_rmvmr.R
\name{pleiotropy_rmvmr}
\alias{pleiotropy_rmvmr}
\title{pleiotropy_rmvmr}
\usage{
pleiotropy_rmvmr(r_input, rmvmr)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{format_rmvmr} function or an object of class \code{MRMVInput} from \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput}}}

\item{rmvmr}{An object containing the output from the \code{ivw_rmvmr} function of class \code{IVW_RMVMR}.}
}
\value{
An object of class \code{"RMVMR_Q"} containing the following components:\describe{
\item{\code{gq}}{A data frame containing the global Q-statistic and p-value after applying a correction for each exposure}
\item{\code{qdat}}{A data frame containing the individual Q-statistic and p-value for each SNP after applying a correction for each exposure}
}
}
\description{
Generates Q-statistics quantifying the degree of heterogeneity in univariate Radial MR analyses applying a correction using the
output from \code{ivw_rmvmr}. The function returns two data frames. The first data frame includes the global Q-statistic for each exposure after applying
a correction, as well as a corresponding p-value. The second data frame contains the individual Q-statistic for each SNP in the corrected univariate
analyses, relative to the exposure given in column \code{exposure}.
}
\examples{
f.data <- format_rmvmr(
    BXGs = rawdat_rmvmr[,c("ldl_beta","hdl_beta","tg_beta")],
    BYG = rawdat_rmvmr$sbp_beta,
    seBXGs = rawdat_rmvmr[,c("ldl_se","hdl_se","tg_se")],
    seBYG = rawdat_rmvmr$sbp_se,
    RSID = rawdat_rmvmr$snp)
rmvmr_output <- ivw_rmvmr(f.data, FALSE)
q_object <- pleiotropy_rmvmr(f.data, rmvmr_output)
q_object$gq
head(q_object$qdat)
}
\references{
Spiller, W., et al., Estimating and visualising multivariable Mendelian randomization analyses within a radial framework. Forthcoming.
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
