% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rmvmr.R
\name{plot_rmvmr}
\alias{plot_rmvmr}
\title{plot_rmvmr}
\usage{
plot_rmvmr(r_input, rmvmr)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{\link{format_rmvmr}} function or an object of class \code{MRMVInput} from \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput}}}

\item{rmvmr}{An object containing the output from the \code{\link{ivw_rmvmr}} function of class \code{IVW_RMVMR}.}
}
\value{
An object of class \code{"RMVMR_plot"} containing the following components:\describe{
\item{\code{p1}}{A radial MVMR plot without correction}
\item{\code{p2}}{A radial MVMR plot with correction}
}
}
\description{
Generates two radial multivariable Mendelian randomization (MVMR) plots. The first plot shows shows the estimated direct effect for each exposure obtained by fitting a radial MVMR model.
Each data point shows the square root weighting for each SNP on the x-axis, and product of the ratio estimate and square root weighting for each SNP on the y-axis. These values are obtained
by performing a univariate radial MR analysis for each exposure using the SNPs displayed, specifically through use of the \code{\link[RadialMR:ivw_radial]{RadialMR::ivw_radial}} function. Only SNPs strongly associated with the corresponding
exposure are used, such that their first stage F-statistic is greater than 10. The second plot applies a correction to each ratio estimate. In both plots, the distance of each observation from the
corresponding regression line is proportional to the contribution of that SNP towards global heterogeneity.
}
\examples{
f.data <- format_rmvmr(
    BXGs = rawdat_rmvmr[,c("ldl_beta","hdl_beta","tg_beta")],
    BYG = rawdat_rmvmr$sbp_beta,
    seBXGs = rawdat_rmvmr[,c("ldl_se","hdl_se","tg_se")],
    seBYG = rawdat_rmvmr$sbp_se,
    RSID = rawdat_rmvmr$snp)
rmvmr_output <- ivw_rmvmr(f.data, FALSE)
plot_object <- plot_rmvmr(f.data, rmvmr_output)
plot_object$p1
plot_object$p2
}
\references{
Spiller, W., et al., Estimating and visualising multivariable Mendelian randomization analyses within a radial framework. Forthcoming.
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
