% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength_rmvmr.R
\name{strength_rmvmr}
\alias{strength_rmvmr}
\title{strength_rmvmr}
\usage{
strength_rmvmr(r_input, gencov)
}
\arguments{
\item{r_input}{A formatted data frame using the \code{\link{format_rmvmr}} function or an object of class \code{MRMVInput} from \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput}}}

\item{gencov}{Calculating heterogeneity statistics using the \code{MVMR} package requires the covariance between the
effect of the genetic variants on each exposure to be known. This can either be estimated from individual level data,
be assumed to be zero, or fixed at zero using non-overlapping samples of each exposure GWAS. A value of 0 is used by default.}
}
\value{
An object of class \code{"S_RMVMR"} containing the following components:\describe{
\item{\code{plot}}{A list containing plots for RMVMR analyses regressing each exposure sequentially upon remaining exposures in the \code{r_input} object. Plots are indexed by the exposure number serving as the outcome for the RMVMR analysis}
\item{\code{qstat}}{A list containing global Q-statistics for RMVMR analyses regressing each exposure sequentially upon remaining exposures in the \code{r_input} object. Indexing follows that of \code{plots} and p-values for global heterogeneity are provided}
\item{\code{qall}}{A list containing the individual Q-statistics and data for RMVMR analyses regressing each exposure sequentially upon remaining exposures in the \code{r_input} object. Indexing follows that of \code{plots}}
}
}
\description{
Calculates Q-statistics quantifying instrument strength. Each exposure is treated as an outcome sequentially, fitting the remaining
exposures within a radial MVMR model. High Q-statistics indicate a high instrument strength, comparable to the Q_x statistic in conventional
MVMR analyses. The function outputs a list of plots, global Q-statistics, and individual Q-contributions indexed by the exposure number ordered
using the \code{\link{format_rmvmr}} function. Named exposures in each list refer to the remaining exposures in the strength RMVMR model.
}
\examples{
f.data <- format_rmvmr(
    BXGs = rawdat_rmvmr[,c("ldl_beta","hdl_beta","tg_beta")],
    BYG = rawdat_rmvmr$sbp_beta,
    seBXGs = rawdat_rmvmr[,c("ldl_se","hdl_se","tg_se")],
    seBYG = rawdat_rmvmr$sbp_se,
    RSID = rawdat_rmvmr$snp)
output <- strength_rmvmr(f.data)

# The following shows the strength plot and Q statistics for exposure 2,
# regressing exposure 2 upon exposures 1 and 3 (which are labeled exposure 1
# and exposure 2 based on ordering in the RMVMR model).

output$plot[[2]]
output$qstat[[2]]
}
\references{
Spiller, W., et al., Estimating and visualising multivariable Mendelian randomization analyses within a radial framework. Forthcoming.
}
\author{
Wes Spiller; Eleanor Sanderson; Jack Bowden.
}
