% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrmvinput_to_rmvmr_format.R
\name{mrmvinput_to_rmvmr_format}
\alias{mrmvinput_to_rmvmr_format}
\title{Convert an object of class MRMVInput from the MendelianRandomization package to the RMVMR rmvmr_format class}
\usage{
mrmvinput_to_rmvmr_format(dat)
}
\arguments{
\item{dat}{Object from \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput}}.}
}
\value{
Object of class \code{rmvmr_format}, the MVMR format
}
\description{
Creates a data.frame with additional class \code{rmvmr_format} from an object of class \code{MRMVInput} generated by \code{\link[MendelianRandomization:mr_mvinput]{MendelianRandomization::mr_mvinput}}.
}
\examples{
bx <- as.matrix(rawdat_rmvmr[,c("ldl_beta", "hdl_beta")])
bxse <- as.matrix(rawdat_rmvmr[,c("ldl_se", "hdl_se")])
dat <- MendelianRandomization::mr_mvinput(bx = bx,
                                          bxse = bxse,
                                          by = rawdat_rmvmr$sbp_beta,
                                          byse = rawdat_rmvmr$sbp_se,
                                          snps = rawdat_rmvmr$snp)
dat <- mrmvinput_to_rmvmr_format(dat)
head(dat)
class(dat)
}
